backend/requirements.txt
fastapi 
uvicorn 
aiomysql 
sqlalchemy 
passlib 
bcrypt==3.2.2 
pydantic-settings 
python-jose 
python-multipart 
redis 
jinja2
cryptography